# docs/api/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HedgeOrderLinkageCreate": {
        "properties": {
          "contract_id": {
            "description": "Linked hedge contract ID",
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "order_id": {
            "description": "Linked order ID",
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "quantity_mt": {
            "description": "Linked quantity in MT",
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "order_id",
          "contract_id",
          "quantity_mt"
        ],
        "title": "HedgeOrderLinkageCreate",
        "type": "object"
      },
      "HedgeOrderLinkageRead": {
        "properties": {
          "contract_id": {
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "quantity_mt": {
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "id",
          "order_id",
          "contract_id",
          "quantity_mt",
          "created_at"
        ],
        "title": "HedgeOrderLinkageRead",
        "type": "object"
      },
...
      "MTMObjectType": {
        "enum": [
          "hedge_contract",
          "order"
        ],
        "title": "MTMObjectType",
        "type": "string"
      },
      "MTMResultResponse": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "entry_price": {
            "title": "Entry Price",
            "type": "string"
          },
          "mtm_value": {
            "title": "Mtm Value",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "$ref": "#/components/schemas/MTMObjectType"
          },
          "price_d1": {
            "title": "Price D1",
            "type": "string"
          },
          "quantity_mt": {
            "title": "Quantity Mt",
            "type": "string"
          }
        },
        "required": [
          "object_type",
          "object_id",
          "as_of_date",
          "mtm_value",
          "price_d1",
          "entry_price",
          "quantity_mt"
        ],
        "title": "MTMResultResponse",
        "type": "object"
      },
      "MTMSnapshotCreate": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "correlation_id": {
            "description": "Caller-provided correlation id for evidence",
            "title": "Correlation Id",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "$ref": "#/components/schemas/MTMObjectType"
          }
        },
        "required": [
          "object_type",
          "object_id",
          "as_of_date",
          "correlation_id"
        ],
        "title": "MTMSnapshotCreate",
        "type": "object"
      },
      "MTMSnapshotResponse": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entry_price": {
            "title": "Entry Price",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mtm_value": {
            "title": "Mtm Value",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "$ref": "#/components/schemas/MTMObjectType"
          },
          "price_d1": {
            "title": "Price D1",
            "type": "string"
          },
          "quantity_mt": {
            "title": "Quantity Mt",
            "type": "string"
          }
        },
        "required": [
          "id",
          "object_type",
          "object_id",
          "as_of_date",
          "mtm_value",
          "price_d1",
          "entry_price",
          "quantity_mt",
          "correlation_id",
          "created_at"
        ],
        "title": "MTMSnapshotResponse",
        "type": "object"
      },
      "OrderPricingConvention": {
        "enum": [
          "AVG",
          "AVGInter",
          "C2R"
        ],
        "title": "OrderPricingConvention",
        "type": "string"
      },
      "OrderRead": {
        "properties": {
          "avg_entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required only for variable orders (USD/MT)",
            "title": "Avg Entry Price"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "price_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriceType"
              }
            ],
            "description": "Fixed or variable pricing"
          },
          "pricing_convention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderPricingConvention"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required only for variable orders (AVG, AVGInter, C2R)"
          },
          "quantity_mt": {
            "description": "Quantity in metric tons (MT)",
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "price_type",
          "quantity_mt",
          "id",
          "order_type",
          "created_at"
        ],
        "title": "OrderRead",
        "type": "object"
      },
      "OrderType": {
        "enum": [
          "SO",
          "PO"
        ],
        "title": "OrderType",
        "type": "string"
      },
...
      "PriceType": {
        "enum": [
          "fixed",
          "variable"
        ],
        "title": "PriceType",
        "type": "string"
      },
      "PurchaseOrderCreate": {
        "properties": {
          "avg_entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required only for variable orders (USD/MT)",
            "title": "Avg Entry Price"
          },
          "price_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriceType"
              }
            ],
            "description": "Fixed or variable pricing"
          },
          "pricing_convention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderPricingConvention"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required only for variable orders (AVG, AVGInter, C2R)"
          },
          "quantity_mt": {
            "description": "Quantity in metric tons (MT)",
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "price_type",
          "quantity_mt"
        ],
        "title": "PurchaseOrderCreate",
        "type": "object"
      },
...
      "SalesOrderCreate": {
        "properties": {
          "avg_entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required only for variable orders (USD/MT)",
            "title": "Avg Entry Price"
          },
          "price_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriceType"
              }
            ],
            "description": "Fixed or variable pricing"
          },
          "pricing_convention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderPricingConvention"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required only for variable orders (AVG, AVGInter, C2R)"
          },
          "quantity_mt": {
            "description": "Quantity in metric tons (MT)",
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "price_type",
          "quantity_mt"
        ],
        "title": "SalesOrderCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hedge Control Platform",
    "version": "0.3.0-phase3-step2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/linkages": {
      "post": {
        "operationId": "create_linkage_linkages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HedgeOrderLinkageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HedgeOrderLinkageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Linkage",
        "tags": [
          "Linkages"
        ]
      }
    },
    "/linkages/{linkage_id}": {
      "get": {
        "operationId": "get_linkage_linkages__linkage_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "linkage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Linkage Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HedgeOrderLinkageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Linkage",
        "tags": [
          "Linkages"
        ]
      }
    },
...
    "/mtm/hedge-contracts/{contract_id}": {
      "get": {
        "operationId": "get_mtm_for_hedge_contract_mtm_hedge_contracts__contract_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contract_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contract Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "As Of Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MTMResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mtm For Hedge Contract",
        "tags": [
          "MTM"
        ]
      }
    },
    "/mtm/orders/{order_id}": {
      "get": {
        "operationId": "get_mtm_for_order_mtm_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "As Of Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MTMResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mtm For Order",
        "tags": [
          "MTM"
        ]
      }
    },
    "/mtm/snapshots": {
      "get": {
        "operationId": "get_mtm_snapshot_mtm_snapshots_get",
        "parameters": [
          {
            "in": "query",
            "name": "object_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MTMObjectType"
            }
          },
          {
            "in": "query",
            "name": "object_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "As Of Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MTMSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mtm Snapshot",
        "tags": [
          "MTM"
        ]
      },
      "post": {
        "operationId": "create_mtm_snapshot_mtm_snapshots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MTMSnapshotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MTMSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Mtm Snapshot",
        "tags": [
          "MTM"
        ]
      }
    },
    "/orders/purchase": {
      "post": {
        "operationId": "create_purchase_order_orders_purchase_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseOrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Purchase Order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/sales": {
      "post": {
        "operationId": "create_sales_order_orders_sales_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Sales Order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{order_id}": {
      "get": {
        "operationId": "get_order_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "Orders"
        ]
      }
    },
...
  }
}